# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server for app.",
        "title": "perna",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/emergency-requests": {
            "get": {
                "description": "Get all Emergency Requests",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EmergencyRequest"
                ],
                "summary": "Get All Emergency Requests",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.EmergencyRequest"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new Emergency Request record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EmergencyRequest"
                ],
                "summary": "Create Emergency Request",
                "parameters": [
                    {
                        "description": "Emergency Request Data",
                        "name": "EmergencyRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interfaces.EmergencyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.EmergencyRequest"
                        }
                    }
                }
            }
        },
        "/api/emergency-requests/{id}": {
            "get": {
                "description": "Get a single Emergency Request by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EmergencyRequest"
                ],
                "summary": "Get Emergency Request by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Emergency Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.EmergencyRequest"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an Emergency Request record by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EmergencyRequest"
                ],
                "summary": "Update Emergency Request by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Emergency Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated Emergency Request Data",
                        "name": "EmergencyRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interfaces.EmergencyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.EmergencyRequest"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an Emergency Request record by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EmergencyRequest"
                ],
                "summary": "Delete Emergency Request by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Emergency Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted Successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/programs": {
            "get": {
                "description": "Get all Programs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Program"
                ],
                "summary": "Get All Programs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.Program"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new Program record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Program"
                ],
                "summary": "Create Program",
                "parameters": [
                    {
                        "description": "Program Data",
                        "name": "Program",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interfaces.Program"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.Program"
                        }
                    }
                }
            }
        },
        "/api/programs/{id}": {
            "get": {
                "description": "Get a single Program by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Program"
                ],
                "summary": "Get Program by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Program ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.Program"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a Program record by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Program"
                ],
                "summary": "Update Program by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Program ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated Program Data",
                        "name": "Program",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/interfaces.Program"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.Program"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a Program record by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Program"
                ],
                "summary": "Delete Program by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Program ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted Successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "db.EmergencyRequest": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "description": "Localização geográfica da vítima (ex: coordenadas GPS)",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "victim_contact": {
                    "description": "Contato da vítima",
                    "type": "string"
                },
                "victim_name": {
                    "description": "Nome da vítima solicitando socorro",
                    "type": "string"
                }
            }
        },
        "db.Program": {
            "type": "object",
            "properties": {
                "contact_email": {
                    "description": "Email de contato",
                    "type": "string"
                },
                "contact_phone": {
                    "description": "Telefone de contato",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "description": "Descrição do programa",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "program_name": {
                    "description": "Nome do programa",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "website_url": {
                    "description": "URL do site do programa",
                    "type": "string"
                }
            }
        },
...
        "interfaces.EmergencyRequest": {
            "description": "Solicitação de socorro com localização da vítima.",
            "type": "object",
            "properties": {
                "location": {
                    "description": "Localização geográfica da vítima (ex: coordenadas GPS)",
                    "type": "string"
                },
                "victim_contact": {
                    "description": "Contato da vítima",
                    "type": "string"
                },
                "victim_name": {
                    "description": "Nome da vítima solicitando socorro",
                    "type": "string"
                }
            }
        },
        "interfaces.Program": {
            "description": "Programas disponíveis para assistência.",
            "type": "object",
            "properties": {
                "contact_email": {
                    "description": "Email de contato",
                    "type": "string"
                },
                "contact_phone": {
                    "description": "Telefone de contato",
                    "type": "string"
                },
                "description": {
                    "description": "Descrição do programa",
                    "type": "string"
                },
                "program_name": {
                    "description": "Nome do programa",
                    "type": "string"
                },
                "website_url": {
                    "description": "URL do site do programa",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}